src/autogen/requirements.txt
-i https://pypi.org/simple
pyyaml==6.0
pyelftools==0.29
